backend/requirements.txt
firebase-admin
redis
google-generativeai
grpcio==1.64.0
pillow
PyPDF2
python-docx
pyyaml